scripts/validator/requirements.txt
pyyaml
jsonschema>=4.0.0
